% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_cube.R
\name{c_cube}
\alias{c_cube}
\title{Instrumental functions computations}
\usage{
c_cube(X_adj, N, DX, r_n)
}
\arguments{
\item{X_adj}{the standardised version of the covariates X}

\item{N}{the size of X}

\item{DX}{the number of covariates}

\item{r_n}{the parameter indexing the number of instrumental function, which is chosen according the the rule used in AS y default.}
}
\value{
a list containing, in order:

  - X_adj he standardised version of the covariates X

  - r_n the parameter indexing the number of instrumental function, which is chosen according the the rule used in AS y default.

  - g_col a vector containing part of the weights

  - Q_AR  a matrix with the weights that enter the statistic T

  - G_X a binary matrix indexing the observations X that fall into the hypercubes indexed by h.
}
\description{
This function defines, for each specified value of r_n the set of indicator funtions h(X_i) which are the key elements for the RE test with co
covariates
}
