% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_fun.R
\name{c_fun}
\alias{c_fun}
\title{Compute the difference between mean of subvectors of two vectors}
\usage{
c_fun(i, i_t, y, z)
}
\arguments{
\item{i}{starting index}

\item{i_t}{final index}

\item{y}{first vector of elements}

\item{z}{second vector of elements}
}
\value{
a real, the difference between means of subvectors of two vectors
}
\description{
Compute the difference between mean of subvectors of two vectors
}
