% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse.R
\name{inverse}
\alias{inverse}
\title{Inverse the function f}
\usage{
inverse(f, lower = -3, upper = 3)
}
\arguments{
\item{f}{the function to be inverted}

\item{lower}{a lower bound for the inverse}

\item{upper}{an lower bound for the inverse}
}
\description{
This function implements the numerical inverse of the function f.
}
