% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S1.R
\name{S1}
\alias{S1}
\title{Core part of the Statistic T}
\usage{
S1(m_bar, sigma_bar, M1, N_k, p)
}
\arguments{
\item{m_bar}{the sample vector of moments for a specified vector $(h_{a,r},y)$}

\item{sigma_bar}{the sample covariance matrix of m_bar}

\item{M1}{number of inequality moments}

\item{N_k}{index of the $ h_{a,r}$ function considered}

\item{p}{parameter p in the statistic}
}
\value{
a real number with the statistic evaluated
}
\description{
This function implements the core part of the Cramer-von-Mises test statistic T, denoted by S in AS.
}
