% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_stat.R
\name{T_stat}
\alias{T_stat}
\title{Computation of the test statistic}
\usage{
T_stat(m_bar, Sigma_bar, prob_weight, N_g, N_k, p)
}
\arguments{
\item{m_bar}{the moments m_bar for the different instrumental functions h considered}

\item{Sigma_bar}{the matrix of all the variances of the moments m_bar for the different instrumental functions h considered}

\item{prob_weight}{vector of weigths for the test statistic}

\item{N_g}{number of instrumental functions h considered}

\item{N_k}{number of moments}

\item{p}{the parameter p in the Statistic.}
}
\value{
a real T which is the test statistic
}
\description{
This function implements the Computation of the test statistic T given in section 3. "Statistical tests" of "Rationalizing Rational Expectations? Tests and Deviations".
}
