% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.min2.R
\name{which.min2}
\alias{which.min2}
\title{Find the min of a list starting from the end}
\usage{
which.min2(x, last.index = FALSE, ...)
}
\arguments{
\item{x}{list of elements}

\item{last.index}{starting from the last index (=TRUE). Default is false}

\item{...}{hypotetical additional elements}
}
\description{
Find the min of a list starting from the end
}
