% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimDev.R
\name{estimDev}
\alias{estimDev}
\title{Estimation of the minimal deviations from rational expectations with unconstrained information set g*}
\usage{
estimDev(psi, y)
}
\arguments{
\item{psi}{vector of subjective expectations}

\item{y}{vector of realisations of an individual outcome.}
}
\description{
This function estimates of the minimal deviations from rational expectations with unconstrained information set. Both vectors should have the same length. If not, one can randomly select a subset of the longer vector with length equal to that of the shorter one. The function returns a function via the approxfun of the package stats. This function can then be evaluated directly on a desired grid.
}
